:art: 프로젝트 구조 부분 변경.
- [X] dto와 mapper는 데이터와 직접적인 관련이 있으므로, domain 디렉토리로 그룹화
