+++ src/main/java/com/ssafy/motif/app/domain/dto/jwt/TokenDto.java
@@ -1,21 +1,18 @@
-package com.ssafy.motif.app.dto.member;
+package com.ssafy.motif.app.domain.dto.jwt;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
 
 @Getter
-@Setter
 @Builder
-@ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class LoginRequestDto {
+public class TokenDto {
 
     private String email;
-    private String password;
+    private String accessToken;
+    private String refreshToken;
 
 }
