+++ src/main/java/com/ssafy/motif/app/domain/dto/member/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.motif.app.dto.member;
+package com.ssafy.motif.app.domain.dto.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
