+++ src/main/java/com/ssafy/motif/app/domain/mapper/PostMapper.java
@@ -1,6 +1,6 @@
-package com.ssafy.motif.app.mapper;
+package com.ssafy.motif.app.domain.mapper;
 
-import com.ssafy.motif.app.dto.post.PostRequestDto;
+import com.ssafy.motif.app.domain.dto.post.PostRequestDto;
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
