--- src/main/java/com/ssafy/motif/app/dto/jwt/TokenDto.java
@@ -1,18 +1,0 @@
-package com.ssafy.motif.app.dto.jwt;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenDto {
-
-    private String email;
-    private String accessToken;
-    private String refreshToken;
-
-}
