--- src/main/java/com/ssafy/motif/app/exception/EmailDuplicateException.java
@@ -1,7 +1,0 @@
-package com.ssafy.motif.app.exception;
-
-public class EmailDuplicateException extends RuntimeException {
-    public EmailDuplicateException(String message) {
-        super(message);
-    }
-}
