--- src/main/java/com/ssafy/motif/app/mapper/ImageMapper.java
@@ -1,9 +1,0 @@
-package com.ssafy.motif.app.mapper;
-
-import com.ssafy.motif.app.dto.image.ProfileImage;
-
-public interface ImageMapper {
-	void profilePicAdd(ProfileImage profileImage);
-	ProfileImage profilePicLoad(String email);
-	void profilePicRemove(String email);
-}
