+++ src/main/java/com/ssafy/motif/app/service/PostService.java
@@ -1,6 +1,6 @@
 package com.ssafy.motif.app.service;
 
-import com.ssafy.motif.app.dto.post.PostRequestDto;
+import com.ssafy.motif.app.domain.dto.post.PostRequestDto;
 import java.util.List;
 
 public interface PostService {
